Address PR #74 round 3: type-system read-only sets, FormatDisplay regression test
- WriteOnlyFormats / ReadOnlyFormats declared as IReadOnlySet<string> so the
  assembly-internal exposure can't accidentally mutate the contents.
- FormatDisplay moved from MtgCsvHelper.BlazorWebAssembly to the core library
  (it has zero Blazor deps), and DisplayNames exposed as internal
  IReadOnlyDictionary so the test project can assert coverage.
- New FormatDisplayTests: every CardMapFactory.Supported format must have a
  display label — would have caught the ARCHIDEKT label bug in 1.3.0.
