+++ MtgCsvHelper/CardMapFactory.cs
@@ -13,9 +13,10 @@ public class CardMapFactory(IConfiguration config, IReferenceCardCatalog catalog
 
 	// Formats whose CSV doesn't carry enough info to populate a complete card without external lookups,
 	// or that don't make sense as targets for our writers. Internal so tests can derive expectations
-	// from the same source of truth as the Readable/Writable filters below.
-	internal static readonly HashSet<string> WriteOnlyFormats = new(StringComparer.OrdinalIgnoreCase) { "CARDKINGDOM" };
-	internal static readonly HashSet<string> ReadOnlyFormats = new(StringComparer.OrdinalIgnoreCase) { "CARDMARKET" };
+	// from the same source of truth as the Readable/Writable filters below. Declared as IReadOnlySet
+	// so the assembly-internal exposure can't accidentally mutate the contents.
+	internal static readonly IReadOnlySet<string> WriteOnlyFormats = new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "CARDKINGDOM" };
+	internal static readonly IReadOnlySet<string> ReadOnlyFormats = new HashSet<string>(StringComparer.OrdinalIgnoreCase) { "CARDMARKET" };
 
 	public static IReadOnlyList<string> ReadableFormats { get; } = [.. Supported.Where(f => !WriteOnlyFormats.Contains(f))];
 	public static IReadOnlyList<string> WritableFormats { get; } = [.. Supported.Where(f => !ReadOnlyFormats.Contains(f))];
